requirements.txt
pandasai==0.5.0
geopandas==0.13.2